[:art: style] ChallengeVC 네트워크 부분 파일 분리
